Linter fixes

+++ test/test_parse_options.cpp
@@ -38,21 +38,21 @@ TEST_CASE("Invalid max-payload", "[options]") {
 
 TEST_CASE("Set all supported options" "[options]") {
   po::variables_map vm;
-  vm.emplace("max-payload", po::variable_value((long) 40000, false));
-  vm.emplace("map-nodes", po::variable_value((int) 1000, false));
-  vm.emplace("map-area",  po::variable_value((double) 0.1, false));
+  vm.emplace("max-payload", po::variable_value(40000L, false));
+  vm.emplace("map-nodes", po::variable_value(1000, false));
+  vm.emplace("map-area",  po::variable_value(0.1, false));
   vm.emplace("changeset-timeout-open", po::variable_value(std::string("10 minutes"), false));
   vm.emplace("changeset-timeout-idle", po::variable_value(std::string("1 hour"), false));
-  vm.emplace("max-changeset-elements", po::variable_value((int) 1000, false));
-  vm.emplace("max-way-nodes", po::variable_value((int) 100, false));
-  vm.emplace("scale", po::variable_value((long) 100, false));
-  vm.emplace("max-way-nodes", po::variable_value((int) 200, false));
-  vm.emplace("max-relation-members", po::variable_value((int) 50, false));
-  vm.emplace("max-element-tags", po::variable_value((int) 10, false));
-  vm.emplace("ratelimit", po::variable_value((long) 1000000, false));
-  vm.emplace("moderator-ratelimit", po::variable_value((long) 10000000, false));
-  vm.emplace("maxdebt", po::variable_value((long) 500, false));
-  vm.emplace("moderator-maxdebt", po::variable_value((long) 1000, false));
+  vm.emplace("max-changeset-elements", po::variable_value(1000, false));
+  vm.emplace("max-way-nodes", po::variable_value(100, false));
+  vm.emplace("scale", po::variable_value(100L, false));
+  vm.emplace("max-way-nodes", po::variable_value(200, false));
+  vm.emplace("max-relation-members", po::variable_value(50, false));
+  vm.emplace("max-element-tags", po::variable_value(10, false));
+  vm.emplace("ratelimit", po::variable_value(1000000L, false));
+  vm.emplace("moderator-ratelimit", po::variable_value(10000000L, false));
+  vm.emplace("maxdebt", po::variable_value(500L, false));
+  vm.emplace("moderator-maxdebt", po::variable_value(1000L, false));
   vm.emplace("ratelimit-upload", po::variable_value(true, false));
   vm.emplace("bbox-size-limit-upload", po::variable_value(true, false));
   REQUIRE_NOTHROW(check_options(vm));
